convert results02 to fluent api now that its working

+++ AtlasSearchNYC/build.sh
@@ -14,16 +14,29 @@ echo "AS: Building container"
 echo
 docker build -t graboskyc/atlassearchfornyc:latest .
 
+EXITCODE=$?
+
 echo 
 echo "AS: Starting container"
 echo
 
-docker stop atlassearchfornyc
-docker rm atlassearchfornyc
-docker run -t -i -d -p 9999:80 --name atlassearchfornyc -e "MDBCONNSTR=${MDBCONNSTR}" --restart unless-stopped graboskyc/atlassearchfornyc:latest
+if [ $EXITCODE -eq 0 ]
+    then
 
-echo
-echo "+================================"
-echo "| END: Atlas Search for .LOCAL NYC"
-echo "+================================"
-echo
+    docker stop atlassearchfornyc
+    docker rm atlassearchfornyc
+    docker run -t -i -d -p 9999:80 --name atlassearchfornyc -e "MDBCONNSTR=${MDBCONNSTR}" --restart unless-stopped graboskyc/atlassearchfornyc:latest
+
+    echo
+    echo "+================================"
+    echo "| END: Atlas Search for .LOCAL NYC"
+    echo "+================================"
+    echo
+
+else
+    echo
+    echo "+================================"
+    echo "| ERROR: Build failed"
+    echo "+================================"
+    echo
+fi
